Build SILVER V2

+++ Projet-WYE/Assets/Scripts/InputEntry.cs
@@ -6,13 +6,13 @@ public class InputEntry
 {
     public bool inGame;
     public string scenarioName;
-    public float endingValue;
+    public int endingValue;
     public string currentPhase;
     public List<string> questionAnsered;
     public string currentAppart;
 
 
-    public InputEntry(bool ingame, Scenario scenario,float value,Phases phase , List<string> questions,string appart)
+    public InputEntry(bool ingame, Scenario scenario,int value,Phases phase , List<string> questions,string appart)
     {
         questionAnsered = new List<string>(); //Init
         inGame = ingame;
